Fix communication timeout integer overflow

+++ src/clientside/config.hpp
@@ -10,7 +10,7 @@ ActuatorMessage build_command_message();
 
 constexpr unsigned long COMMAND_MESSAGE_INTERVAL_MS = 100;
 constexpr unsigned long COMMUNICATION_RESET_MS = 50;
-constexpr uint8_t COMMUNICATION_TIMEOUT_S = 3;
+constexpr uint16_t COMMUNICATION_TIMEOUT_S = 3;
 
 constexpr uint16_t BATT_SCALE_PRE_OFFSET = -11;
 constexpr uint16_t BATT_SCALE_NUM = 1;
