+++ src/towerside/config.hpp
@@ -11,7 +11,7 @@ void apply(const ActuatorMessage &command);
 ActuatorMessage build_safe_state(const ActuatorMessage &current_state);
 SensorMessage build_sensor_message();
 
-constexpr uint8_t COMMUNICATION_TIMEOUT_S = 10; // Go to safe state after this many seconds without contact
+constexpr uint16_t COMMUNICATION_TIMEOUT_S = 10; // Go to safe state after this many seconds without contact
 constexpr unsigned long SENSOR_MSG_INTERVAL_MS = 100; // Rate to send sensor messages at
 constexpr unsigned long COMMUNICATION_RESET_MS = 50; // maximum time between successive characters in the same message
